+++ src/builder.h
@@ -5,6 +5,6 @@
 
 #define SMALL_BOARD_NB_CELLS 4
 
-int builder_create_small_board(s_board *b, int nb_player);
+int builder_create_small_board(s_board *b, uint8_t nb_player);
 
 #endif
